[ALPS07401732] apusys: mt6985: refine power off check
Bypass the power cycle if got SAPU lock ipi, this is to prevent
from getting false alarm of power off timeout.

This patch is to prevent the below case.
When SAPU locked the RV and the RV will not enter deep idle.
If a normal ipi is sent from APMCU, it will trigger the power
off timeout check and cause a false alarm.

So just bypass all the power off timeout check on the same
power on cycle of the SAPU cases until it is power off.

MTK-Commit-Id: 5abc8ce67f997e502a036e270d4f2e04e5b00bb2

Signed-off-by: Andy Teng <andy.teng@mediatek.com>
CR-Id: ALPS07401732
Feature: apu_microp_kernel
Change-Id: I434435574ba382c31336a9d26759964f5f82d03e