[ALPS07461400] Touch: KE at goodix_fw_update_thread
KE Cause:during the execution, the value of the global variable "mode"
changes,which satisfies condition (fwu->mode & UPDATE_MODE_SRC_SYSFS)
and enter the if structure. At this time,the value of "fw_sysfs" is
NULL,resulting in KE.
KE Solution: define a local variable in the function and pass the value
of "mode" to the local variable. No change in the value of the global
variable will affect the execution result of this function.

MTK-Commit-Id: 1ff115e2673dcdb0901bbc5ed7811b5055fd88c7

Signed-off-by: ot_minjie.li <ot_minjie.li@mediatek.com>
CR-Id: ALPS07461400
Feature: touch_driver
Change-Id: Ia966814059a8210dc42ecd60096f332c2a1e72a5