[ALPS07490921] PCI: mediatek-gen3: Control BBCK2 through clkbuf API
BBCK2 is clkbuf 26M. Because of the performance requirements
of WiFi, the normal situation requires always on.
There are two control modes of BBCK2. One is software control.
The other is its own hardware mode. It will be turned off when
no one requests. When PCIe is on, PCIe needs software enable bbck2.

MTK-Commit-Id: f75bb1b0fc6a6cb2fd7dca6ccb68fff652e8373e

CR-Id: ALPS07490921
Feature: pcie
Signed-off-by: qizhong cheng <qizhong.cheng@mediatek.com>
Change-Id: I5ddc81a51cafb4f73af28c69110ec7189b1bd2ed