[ALPS07778378] mmqos: set channel bw by vmmrc
1. DPC HW will use sideband on/off signal to notify VMMRC to change the
on/off table, and then the mminfra frequency will changed.
So the display channel bw should only be set on the "on table" in VMMRC,
mmqos should record the display channel bw. And then minus display bw
when setting to "off table". The bw write to VMMRC should change to
16MB/s a unit.
Also add mmqos UT code for channel bw.
2. Ignore display common port for calculating emi bw if DPC_ENALBED.
3. TODO: add back read, write VMMRC HW.

MTK-Commit-Id: 96cb99add7665e409b3698308765270391d460c1

Signed-off-by: wendy-st.lin <wendy-st.lin@mediatek.com>
CR-Id: ALPS07778378
Feature: MMQoS
Change-Id: I8bb68d80b374ae8594ce4fc9a0f184ed599a3345