[ALPS08029113] DMA-BUF: fix kcsan kernel api dump db
1. The debug_alloc_size of mtk_heap_debug is multi-threaded and causes
KCSAN Kernel API dump. This is used for debugging and is currently not
used, and debug_alloc_size is removed.

2. Refilling is also accessed by multiple threads to generate KCSAN
Kernel API dump, but it will not affect the logic. Whether refill is
needed is ultimately judged according to the pool size. The purpose of
adding refill is to reduce the number of calls to wake_up_process by
alloc_largest_available in mtk_system_heap.
Refer to guangyes wiki:
https://wiki.mediatek.inc/pages/viewpage.action?pageId=1218353474,
use __no_kcsan to fix it.

MTK-Commit-Id: 33f71b9368adaed555f1cbba03341362dc755910

Signed-off-by: jianjiao zeng <jianjiao.zeng@mediatek.com>
CR-Id: ALPS08029113
Feature: dma_buf_kernel
Change-Id: I8db0e6a8980e36e114704894c31f0788ae48860e