[ALPS08609695] arm-smmu-v3: Set TTL invalidation hint better
When io-pgtable unmaps a whole table, rather than waste time walking it
to find the leaf entries to invalidate exactly, it simply expects
.tlb_flush_walk with nominal last-level granularity to invalidate any
leaf entries at higher intermediate levels as well. This works fine with
page-based invalidation, but with range commands we need to be careful
with the TTL hint - unconditionally setting it based on the given level
3 granule means that an invalidation for a level 1 table would strictly
not be required to affect level 2 block entries. Its easy to comply
with the expected behaviour by simply not setting the TTL hint for
non-leaf invalidations, so lets do that.

Signed-off-by: Robin Murphy <robin.murphy@arm.com>
Link: https://lore.kernel.org/r/b409d9a17c52dc0db51faee91d92737bb7975f5b.1685637456.git.robin.murphy@arm.com
Signed-off-by: Will Deacon <will@kernel.org>

MTK-Commit-Id: 9384151901de1594388e4387cd122291c7de8062

Signed-off-by: Yunfei Wang <yf.wang@mediatek.com>
CR-Id: ALPS08609695
Feature: iommu_kernel
Change-Id: I4999b56a0352f90cb7d9eb4e532b318b2db8117d