[ALPS08576950] selinux: Add back the checking of enforcing
We used to use `enforcing_enabled()` to check if SELinux is in
permissive mode. However, in kernel-6.6, this function requires an
unexported variable, which causes external modules to be unable to use
this function. Therefore, in line with the original design, we now use
the value of `sad->result` to determine the permissive mode.

MTK-Commit-Id: 4d7f056b6f88b453f80c7c0210f2453fac70dd63

Fixes: 36ecef7ddcb8 ("[ALPS08309937] selinux_warning: fix build error")
Signed-off-by: Cheng-Jui Wang <cheng-jui.wang@mediatek.com>
CR-Id: ALPS08576950
Feature: selinux
Change-Id: Iec3cb79a8d00e7daadc27e90d2db0ae6cb2ac7e0