[ALPS08696318]  MSDC:modify RA for better performance
background:
    RA can try to put the data on the disk into the pagecache first,
thereby improving the read and write performance of the file system.
For better SD card read and write performance, we will set an
appropriate RA value in the init rc. However, if the card is inserted
after booting, the RA value cannot be modified.
question:
    Due to GKI limitations, we cannot modify RA directly at the core
layer.
solution:
    This patch is used in our underlying drievr to access and modify
the RA value of the block device of the upper layer bdi. We will
modify the appropriate value when the RA is relatively small and the
upper layer has prepared the bdi. In this way, good read and write
performance can be ensured even after the card is inserted after the
computer is turned on.
Test:
1.Insert/remove normal,bad card test ok.
2,sd r/w stress test ok.
3,sd read_ahead size test ok.
4,sd read peroformance test ok.

MTK-Commit-Id: 0a81a10d0642691e3fd50280dc5409ef86424279

Signed-off-by: mtk28646 <wey.zhang@mediatek.com>
CR-Id: ALPS08696318
Feature: MSDC
Change-Id: I0c5368ba326502dc72ab4090d944a045553ecb75