[ALPS08659114] xhci: sync upstream patch 2024_02_18-46
[Description]
usb: xhci: xhci-ring: Use sysdev for mapping bounce buffer

As mentioned in:
  commit 474ed23a6257 ("xhci: align the last trb before link if it is
easily splittable.")

A bounce buffer is utilized for ensuring that transfers that span across
ring segments are aligned to the EPs max packet size.  However, the device
that is used to map the DMA buffer to is currently using the XHCI HCD,
which does not carry any DMA operations in certain configrations.
Migration to using the sysdev entry was introduced for DWC3 based
implementations where the IOMMU operations are present.

Replace the reference to the controller device to sysdev instead.  This
allows the bounce buffer to be properly mapped to any implementations that
have an IOMMU involved.

cc: stable@vger.kernel.org
Fixes: 4c39d4b949d3 ("usb: xhci: use bus->sysdev for DMA configuration")
Signed-off-by: Wesley Cheng <quic_wcheng@quicinc.com>
Signed-off-by: Mathias Nyman <mathias.nyman@linux.intel.com>
Link: https://lore.kernel.org/r/20230915143108.1532163-2-mathias.nyman@linux.intel.com
Signed-off-by: Greg Kroah-Hartman <gregkh@linuxfoundation.org>

[Test]
USB Host

MTK-Commit-Id: 9f322b246f7a6d9282d123fe7829fe180a8e0038

Signed-off-by: yu-chen.liu <yu-chen.liu@mediatek.com>
CR-Id: ALPS08659114
Feature: usb_ssusb
Change-Id: I61ed1858a982d8d671b796ecf681f18364fa738b