[ALPS08659114] xhci: sync upstream patch 2024_02_18-50
[Description]
usb: host: xhci: Avoid XHCI resume delay if SSUSB device is not present

There is a 120ms delay implemented for allowing the XHCI host controller to
detect a U3 wakeup pulse.  The intention is to wait for the device to retry
the wakeup event if the USB3 PORTSC doesnt reflect the RESUME link status
by the time it is checked.  As per the USB3 specification:

  tU3WakeupRetryDelay ("Table 7-12. LTSSM State Transition Timeouts")

This would allow the XHCI resume sequence to determine if the root hub
needs to be also resumed.  However, in case there is no device connected,
or if there is only a HSUSB device connected, this delay would still affect
the overall resume timing.

Since this delay is solely for detecting U3 wake events (USB3 specific)
then ignore this delay for the disconnected case and the HSUSB connected
only case.

[skip helper function, rename usb3_connected variable -Mathias ]

Signed-off-by: Wesley Cheng <quic_wcheng@quicinc.com>
Signed-off-by: Mathias Nyman <mathias.nyman@linux.intel.com>
Link: https://lore.kernel.org/r/20231019102924.2797346-20-mathias.nyman@linux.intel.com
Signed-off-by: Greg Kroah-Hartman <gregkh@linuxfoundation.org>
Signed-off-by: Sasha Levin <sashal@kernel.org>

[Test]
USB Host

MTK-Commit-Id: 279701bff2a8c6526d5efb98338d89d759dc15fb

Signed-off-by: yu-chen.liu <yu-chen.liu@mediatek.com>
CR-Id: ALPS08659114
Feature: usb_ssusb
Change-Id: I2c69e81da5fb5611730453b7a06264916303b4f4