[ALPS08709396] iommu: convert to platform remove callback return void
[Description]
The .remove() callback for a platform driver returns an int
which makes many driver authors wrongly assume its possible
to do error handling by returning an error code. However the
value returned is (mostly) ignored and this typically results
in resource leaks. To improve here there is a quest to make
the remove callback return void. In the first step of this
quest all drivers are converted to .remove_new() which already
returns void.

Trivially convert this driver from always returning zero in
the remove callback to the void returning variant.

Signed-off-by: Uwe Kleine-K?nig <u.kleine-koenig@pengutronix.de>
Reviewed-by: AngeloGioacchino Del Regno <angelogioacchino.delregno@collabora.com>
Link: https://lore.kernel.org/r/20230321084125.337021-8-u.kleine-koenig@pengutronix.de
Signed-off-by: Joerg Roedel <jroedel@suse.de>

[Test]
check whether it can boot to home screen normally.

MTK-Commit-Id: 40af0969e73f675c720c6d757228f817ae398bb7

CR-Id: ALPS08709396
Signed-off-by: Qiming Sun <qiming.sun@mediatek.com>
Feature: iommu_kernel
Change-Id: I03e2e3d0225a38c062b7d4e26a075b34b8b3ea80