[ALPS08709396] iommu: lookup phandle to retrieve syscon to infracfg
[Description]
This driver will get support for more SoCs and the list of infracfg
compatibles is expected to grow: in order to prevent getting this
situation out of control and see a long list of compatible strings,
add support to retrieve a handle to infracfgs regmap through a
new "mediatek,infracfg" phandle.

In order to keep retrocompatibility with older devicetrees, the old
way is kept in place.

Signed-off-by: AngeloGioacchino Del Regno <angelogioacchino.delregno@collabora.com>
Reviewed-by: Miles Chen <miles.chen@mediatek.com>
Reviewed-by: Yong Wu <yong.wu@mediatek.com>
Reviewed-by: Matthias Brugger <matthias.bgg@gmail.com>
Link: https://lore.kernel.org/r/20220616110830.26037-3-angelogioacchino.delregno@collabora.com
Signed-off-by: Joerg Roedel <jroedel@suse.de>

[Test]
Check whether it can boot to home screen normally.

MTK-Commit-Id: af5615c7404a66bd74eb0d4f363154ec77cfb65e

CR-Id: ALPS08709396
Signed-off-by: Qiming Sun <qiming.sun@mediatek.com>
Feature: iommu_kernel
Change-Id: I2362b7c02018f63d08e49da90243f57603dd7e38