[ALPS08709396] iommu: get regionid from larb/port id
[Description]
The dma-ranges is not allowed for dts leaf node. But we still would
like to separate to different masters into different iova regions.

Thus we have to separate it by the HW larbid and portid. For example,
larb1/2 are in region2 and larb3 is in region3. The problem is that
some ports inside a larb are in region4 while some ports inside this
larb are in region5. Therefore I define a "iova_region_larb_msk" to help
record the information for each a port. Take a example for a larb:
 [1] = ~0: means all ports in this larb are in region1;
 [2] = BIT(3) | BIT(4): means port3/4 in this larb are region2;
 [3] = ~(BIT(3) | BIT(4)): means all the other ports except port3/4
                           in this larb are region3.

This method also avoids the users forget/abuse the iova regions.

Signed-off-by: Yong Wu <yong.wu@mediatek.com>
Reviewed-by: AngeloGioacchino Del Regno <angelogioacchino.delregno@collabora.com>
Link: https://lore.kernel.org/r/20230411093144.2690-5-yong.wu@mediatek.com
Signed-off-by: Joerg Roedel <jroedel@suse.de>

[Test]
Since the change affect attach flow ,so check whether it can
attach dev normally.

MTK-Commit-Id: f56c54854f00bae68cdf5b7743fd39bb0c6513d4

CR-Id: ALPS08709396
Signed-off-by: Qiming Sun <qiming.sun@mediatek.com>
Feature: iommu_kernel
Change-Id: I56321226dc659085f4cd5e71e678cfe964705ab9