[ALPS08694711] arm64: dts: mt6833: modify dma-ranges
[Description]
Kernel6.6 obtains the dma-ranges attribute logic adjustment,
which is obtained from the parent node by default, and the
dma-ranges configured on this node are invalid. Therefore,
it is necessary to move the nodes whose dma-ranges attribute
exceeds 4G to the soc, and configure the soc dma-ranges
to be 0~16G.

[Test]
iommu basic test pass.

MTK-Commit-Id: 8b022fde484324dd4aa42ee1eef88a0809e07aac

Change-Id: Ifab36b5de1b5bca5d309621f4e7573e8ca957182
CR-Id: ALPS08694711
Signed-off-by: "ot_vikram.b" <ot_vikram.b@mediatek.com>
Feature: iommu_kernel