[ALPS08848007] irq_mon: improve false alarm detection
[Description]
Currently, the IRQ Monitor will label functions exceeding the time
threshold as false alarms. However, if a function contributes 99% of
the time without being flagged as a false alarm, and another piece of
code contributes the remaining 1% triggering a warning, this scenario
should also be classified as a false alarm.
So, this patch change IRQ Monitor to consider it as false alarm when
it exceeds more than half of the time.

[Test]
PDAS

MTK-Commit-Id: d2de04b49126882473daf62c13a2925b2de93bce

Signed-off-by: Cheng-Jui Wang <cheng-jui.wang@mediatek.com>
CR-Id: ALPS08848007
Feature: IRQ Monitor
Change-Id: If921f1ea9070b01d4c8d3d2f37e58f10c6653383