[ALPS08861160] xhci: sync upstream patch 2024/03/18-11
[Description]
UPSTREAM: xhci: add support to allocate several interrupters

Modify the XHCI drivers to accommodate for handling multiple event rings in
case there are multiple interrupters.  Add the required APIs so clients are
able to allocate/request for an interrupter ring, and pass this information
back to the client driver.  This allows for users to handle the resource
accordingly, such as passing the event ring base address to an audio DSP.
There is no actual support for multiple MSI/MSI-X vectors.

[Test]
usb host

MTK-Commit-Id: 9193adccd900e3976cfcc5a589bb8cf0c4723802

Signed-off-by: yu-chen.liu <yu-chen.liu@mediatek.com>
CR-Id: ALPS08861160
Feature: usb_ssusb
Change-Id: Ife7fd440d88e0a0b2e1924a48dc7fa6181a717d5