[ALPS08861160] xhci: sync upstream patch 2024/03/18-20
[Description]
FROMGIT: xhci: update event ring dequeue pointer position to
controller correctly

The event ring dequeue pointer field (ERDP) in xHC hardware is
used to inform controller how far the driver has processed
events on the event ring.

In the case all events are handled and event ring is empty then
the address of the TRB after the last processed one should be
written. This TRB is both the enqueue and dequeue pointer.

But in case we are writing the ERDP in the middle of processing
several events then ERDP field should be written with the "up
to and including" address of the last handled event TRB.

Currenly each ERDP write by driver is done as if all events
are handled and ring is empty.

Fix this by adjusting the order when software dequeue
"inc_deq()" is called and hardware dequeue
"xhci_update_erst_dequeue()" is updated.

[Test]
usb host

MTK-Commit-Id: b3b01f4bbcc61da045a09c392a5d16fa1b20f58b

Signed-off-by: yu-chen.liu <yu-chen.liu@mediatek.com>
CR-Id: ALPS08861160
Feature: usb_ssusb
Change-Id: I8b51a0325c30f2994cd029a9aa474141d20500dd