[ALPS08963419] arm64: dts: mt6833: modify soc node
[Description]
Kernel6.1 obtains the dma-ranges attribute logic adjustment,
which is obtained from the parent node by default, and the
dma-ranges configured on this node are invalid. Therefore,
it is necessary to move the nodes whose dma-ranges attribute
exceeds 4G to the soc, and configure the soc dma-ranges to be
0~16G.

[Test]
iommu basic test pass.

MTK-Commit-Id: 703b90bc2384d29195e84f6a0ff624c983573b2d

Change-Id: Ieb66ffa36512b5c60f442361e36e4167bb228080
CR-Id: ALPS08963419
Signed-off-by: "ot_vikram.b" <ot_vikram.b@mediatek.com>
Feature: iommu_kernel