[ALPS08883069] drivers: fix crash issue without nbl_trace node[1/1]
[Description]
In some case the nbl_trace reserve memory node will not exist which
 will cause mbraink insmod fail. Systemd will reload mbraink when
load mbraink_v6991(depends on mbraink) and it will cause resource
conflict with the first stage of load mbraink.
So we remove the return fail of no nbl_trace case in temp.

[Test]
1.collect vcpu ok with nbl_trace node in hyp.
2.mbraink insmod success and boot up ok without nbl_trace node
in yocto. collect vcpu will print expected log and not crash.

MTK-Commit-Id: 924a735da43cdbe8f76af5a39dd7a6d23b28eb89

Feature: MBrain
Signed-off-by: Dengjun.Su <dengjun.su@mediatek.com>
CR-Id: ALPS08883069
Change-Id: Ie5841831f8dce4c90222e4d35eb16b494c692444