[ALPS08564265] Touch: fix touch issue
[Description]
If the PID of the firmware being updated does not match the PID of the
firmware currently in the touch IC, it may result in the firmware
being unable to revert back. When attempting to do firmware update
again, an error may occur.
Therefore, if the PID does not match, the firmware should not be
updated, and this situation is considered a firmware update failure.

[Test]
1. touch ok for 1.5k panel and nt36672e using the GT9895 touch IC.
2. the firmware is not updated if PID mismatches are detected through
the logs for different loads.

MTK-Commit-Id: 14443a67054493aeffa306940279e706f89d61b8

Signed-off-by: mtk27743 <jianming.liu@mediatek.com>
CR-Id: ALPS08564265
Feature: touch_driver
Change-Id: I36260be3ffb3c4d01aaf1837088a7465e9378b35