[ALPS08958911] SPI:add cs_setup_cnt default val
[Description]
1. setup_cs_cnt needs to be configured by the customer. If the
user does not configure it, this code is used as a defense measure
to make set_cnt have a correct value.
2. If the customer does not configure it, the previous cs_cnt value
is 0 the rising edge of clk and MO data will be printed out together,
which may cause data abnormally.
2. According to DEs suggestion, the default value of cs_setup_cnt
should be greater than or equal to the value of high_cnt.
3. In the bottom layer, drv sets this default value before each
transfer.
4. Because there are two usage methods for spi io, spi_sync and
mem_op. the original value of  users cs_setup_cnt cannot be
obtained through cur_msg, so the previous setting value can only
be viewed through the register value.

[Test]
1,spi stress test UT.

MTK-Commit-Id: 483a5021a068430381a89952e4f6e9d5c32198e0

Signed-off-by: yijian jia <yijian.jia@mediatek.com>
CR-Id: ALPS08958911
Feature: spi_master
Change-Id: Ia8560ea30102eeabf4fae4f92f5c567ee1d12da5