[ALPS09205097] arm64: dts: mt6899: usnd add semaphore
[Description]
On the new platform, a new IRQ clean register is used. However,
since the method to clean ultrasound IRQ is common code, it is
necessary to use the toggle clean method corresponding to the new
register on the mt6878 platform.

[Test]
Ultrasound test, audio stress, vow and ultrasound.

MTK-Commit-Id: 1e2f5a9ccdfa70db13a51ae53da52b2cb656e7ba

CR-Id: ALPS09205097
Feature: Ultrasound Proximity Sensor
Change-Id: I0681279a5e09bd5f89532b908ebfcc19b30e727d
Signed-off-by: Min Liu <min.liu@mediatek.com>