[ALPS09246989] KPM: xhci: Add a quirk for writing ERST in high-low order
[Description]
This quirk is for the controller that has a limitation in supporting
separate ERSTBA_HI and ERSTBA_LO programming. Its supported when
the ERSTBA is programmed ERSTBA_HI before ERSTBA_LO. Thats because
the internal initialization of event ring fetches the
"Event Ring Segment Table Entry" based on the indication of ERSTBA_LO
written.

[Test]
usb host

MTK-Commit-Id: bfc6cfcbdb4e6cadabb41112447422a9b0881537

CR-Id: ALPS09246989
Feature: usb_ssusb
Bug: 346655675
Link: https://lore.kernel.org/lkml/1718019553-111939-3-git-send-email-dh10.jung@samsung.com/
Change-Id: I24d4d479aadb18786f2df97389288855fe32f23d
Signed-off-by: Daehwan Jung <dh10.jung@samsung.com>