[ALPS09047177] xhci: mbrain: fix the un-changed state case
[Description]
if some device keeps in a state over 2s, we assume it is in the final
state (enumeration ok or failed).
However for the enumeration failed case, xhci won’t send the urb again
(xhci_trace_ep_urb_giveback() won’t be invoked). And won’t update the
mbrain db. To fix that, queue a work to handle it.

[Test]
insert an usb device and update mbrain db well

MTK-Commit-Id: fd56b09f0b200dc90f0d66ba78e8eedf42626151

CR-Id: ALPS09047177
Feature: usb_ssusb
Change-Id: I6a98a1ebac96f07f95239086047d041a1814092d
Signed-off-by: Denis Hsu <denis.hsu@mediatek.com>