[ALPS06513286] regulator: mt6319: Support VMM regulator
VMM regulator is a kind of regulator controls buck of VMM.
There are two methods to control voltage, controlled by APMCU or
camera computing unit. Due to the fact that CCU is not ready.
We use APMCU to control DVFS (Dynamic voltage frequency scaling)
at the start of project.

MTK-Commit-Id: 63744d729f7b812520e523bddf1074140b6c92df

CR-Id: ALPS06513286
Feature: Regulator(regulator)
Signed-off-by: mtk13668 <yuan-jung.kuo@mediatek.com>
Change-Id: I76e22b0190e099ce5a294634c6571fe60849de1a