[ALPS07249242] media: platform: cam: sof and deque thread race
When SOF isr thread and deque worker are running at the same time,
media request may be free in the middle of isr thread because user
grabs the frame on vb2_done and makes request ref cnt to 0.

Then it trigger media request release operation which frees
the request in use.

Add a mutex lock in isr state handler to prevent it.

MTK-Commit-Id: db48d627a363b80712e23b762c386afbb0237279

CR-Id: ALPS07249242
Feature: [Android Default] Camera Application Basic Functions
Signed-off-by: Ryan Yu <ryan.yu@mediatek.com>
Change-Id: I00e4cbcd101d5c6bf43d560e9964459e45960727