[ALPS07782929] usb: rndis: sync patch 1
commit 8ae01239609b
("usb: gadget: u_ether: fix a potential null pointer dereference")

f_ncm tx timeout can call us with null skb to flush
a pending frame.  In this case skb is NULL to begin
with but ceases to be null after dev->wrap() completes.

In such a case in->maxpacket will be read, even though
weve failed to check that in is not NULL.

Though Ive never observed this fail in practice,
however the flush operation simply does not make sense with
a null usb IN endpoint - theres nowhere to flush to...
(note that were the gadget/device, and IN is from the point
 of view of the host, so here IN actually means outbound...)

Cc: Brooke Basile <brookebasile@gmail.com>
Cc: "Bryan ODonoghue" <bryan.odonoghue@linaro.org>
Cc: Felipe Balbi <balbi@kernel.org>
Cc: Greg Kroah-Hartman <gregkh@linuxfoundation.org>
Cc: Lorenzo Colitti <lorenzo@google.com>
Signed-off-by: Maciej Żenczykowski <maze@google.com>
Link: https://lore.kernel.org/r/20210701114834.884597-6-zenczykowski@gmail.com
Signed-off-by: Greg Kroah-Hartman <gregkh@linuxfoundation.org>

MTK-Commit-Id: 3a0daecdcf6a76fc583b02be10d10edf08dcafe3

Signed-off-by: Ken Peng <ken.peng@mediatek.com>
CR-Id: ALPS07782929
Feature: usb_ssusb
Change-Id: If3fc4ffd51c486ff17285b6cf1a6fb4cee8df809