[ALPS07936658] media: cam: raw switch: removed apply_raw_switch
This patch is to remove apply_raw_switch from job ops. The
reason is that we want to reuse the common flow of stream on,
such as initialize_engines().

In this change, I put the initialize_opt in job so that we can
reuse the original initialize_engines() function and pass the
original initialize_opt as the parameters.

MTK-Commit-Id: 4425db5aff16c39b77585eff7a98e57d5477c6f5

CR-Id: ALPS07936658
Feature: [Android Default] Camera Application Basic Functions
Change-Id: Ie94aff308674b129ca61a4ed47a2548dcfd16204
Signed-off-by: Frederic Chen <frederic.chen@mediatek.com>