[ALPS08137807] ANDROID: preserve CRC for xhci symbols
In release 6.1.24 a new .h file was included in the xhci code, which
caused the CRCs to change as some structures changed into "real"
structures instead of anonymous definitions.  So preserve the CRCs by
commenting out the #include if GENKSYMS is being calculated.

This will be removed the next KABI break, as it shouldnt be sticking
around long.

MTK-Commit-Id: 2daa35d96037aee22ccf4dae4f8e1e1003d38c9e

Bug: 161946584
Fixes: 167c05646f67 ("xhci: also avoid the XHCI_ZERO_64B_REGS quirk with a passthrough iommu")
Signed-off-by: Greg Kroah-Hartman <gregkh@google.com>
Change-Id: I59838b0af869d3e17fc73d72eb473190c50281fd
Signed-off-by: Chih-Chun Wang <Chih-Chun.Wang@mediatek.com>
CR-Id: ALPS08137807
Feature: usb_ssusb