[ALPS07820708] ufs: mediatek: add ddren setting
Control of DDR_EN on MT6989 and later platforms has been switched from
SPM to EMI. To avoid abnormal access to DRAM, it is necessary to wait
for `ddren_ack` or assert `ddren_urgent` after sending `ddren_req`.

This patch adds the DDR_EN configuration in the UFS init flow, which
uses the method of asserting `ddren_urgent` to maintain performance.

MTK-Commit-Id: 1739ea9bcd2788ac99f603b402bbf215f5922c3a

Signed-off-by: Naomi Chu <naomi.chu@mediatek.com>
CR-Id: ALPS07820708
Feature: ufs
Change-Id: If13a772cdc2ab574f95fba7b5cb198bfe2f28422