[ALPS08162414] ccci: remove log from spin lock
[detail]
1.It is not safe to use printk in the spin_lock_irqsave function,
because this function will disable interrupts and obtain a spin lock,
which may cause other interrupts to fail to respond in time,
thus affecting the stability of the system.
2.The printk function itself will acquire a lock to ensure the atomicity
of the output. If printk is used in the interrupt handler,
it may cause a deadlock, resulting in a system crash.

[solution]
Remove log from spin lock and print log after spin unlock.

MTK-Commit-Id: 06535d833b7efe3095a124dd4568b3a533cf300c

Signed-off-by: Weiqiang Hao <weiqiang.hao@mediatek.com>
CR-Id: ALPS08162414
Feature: CCCI_HIF_CONTROL
Change-Id: Ia308d22c1279f27e9d945dafb9a9c6840276cc1b