[ALPS08181744] DISP: fix bwm dual pipe issue
[Detail]
Fix bwm dual pipe issue, as unchanged layer
was not reset flag when layer content not change,
but dst x is changed, some content will disply on
right pipe, so when get ratio from right pipe, the
frame is the first frame. The ratio will be zero
or random value.

MTK-Commit-Id: 3aa6c35c1a913e7496f19ef16952a53c81b4c175

Signed-off-by: yonggang.yu <yonggang.yu@mediatek.com>
CR-Id: ALPS08181744
Feature: display_mobile
Change-Id: I2a9abc4fdae1967ba55e2fac3c5382ad4cd0b590