[ALPS08375446] Do not free ultrasound reserve DRAM
[Situation]
The AFEs SRAM is exhausted,and the ultrasound uses the
reserved DRAM as the data path. when the ultrasound is
turned off,scp_ultra_enable will become false, but the AFE
HW free may not be over yet, and it will go to the process
of releasing the reserved DRAM, causeing KE to occur.
[Solution]
When the AFE HW is free, it is judged whether the DRAM of
substream is in the ultra reserved memory. if it is, the
release memory process will be skipped.

MTK-Commit-Id: 504050540e96791985b57c91e0de04583381176e

Signed-off-by: Min Liu <min.liu@mediatek.com>
CR-Id: ALPS08375446
Feature: Ultrasound Proximity Sensor
Change-Id: I221607234afab4699916226f58100d668438129e