[ALPS08609695] arm-smmu-v3: Convert to remove callback returning void
The .remove() callback for a platform driver returns an int which makes
many driver authors wrongly assume its possible to do error handling by
returning an error code. However the value returned is (mostly) ignored
and this typically results in resource leaks. To improve here there is a
quest to make the remove callback return void. In the first step of this
quest all drivers are converted to .remove_new() which already returns
void.

Trivially convert this driver from always returning zero in the remove
callback to the void returning variant.

Signed-off-by: Uwe Kleine-König <u.kleine-koenig@pengutronix.de>
Link: https://lore.kernel.org/r/20230321084125.337021-4-u.kleine-koenig@pengutronix.de
Signed-off-by: Joerg Roedel <jroedel@suse.de>

MTK-Commit-Id: 5f5825eee0feadf097eb6fea049d7a8680451abb

Signed-off-by: Yunfei Wang <yf.wang@mediatek.com>
CR-Id: ALPS08609695
Feature: iommu_kernel
Change-Id: Ia4e86f3d94026f515359fb1ed119b288b7dfa3a1