[ALPS08609695] arm-smmu-v3: Add explicit feature for nesting
In certain cases we may want to refuse to allow nested translation even
when both stages are implemented, so lets add an explicit feature for
nesting support which we can control in its own right. For now this
merely serves as documentation, but it means a nice convenient check
will be ready and waiting for the future nesting code.

Signed-off-by: Robin Murphy <robin.murphy@arm.com>
Reviewed-by: Nicolin Chen <nicolinc@nvidia.com>
Link: https://lore.kernel.org/r/136c3f4a3a84cc14a5a1978ace57dfd3ed67b688.1683731256.git.robin.murphy@arm.com
Signed-off-by: Will Deacon <will@kernel.org>

MTK-Commit-Id: 5cebf9348917ab6881138091a03eb7be75d6d9ea

Signed-off-by: Yunfei Wang <yf.wang@mediatek.com>
CR-Id: ALPS08609695
Feature: iommu_kernel
Change-Id: I448b99101dd1c314600ecd1e7d0b5348d8996a91