[ALPS08609695] arm-smmu-v3: Change vmid alloc strategy from bitmap to ida
For current implementation of vmid allocation of arm smmu-v3, a per-smmu
devide bitmap of 64K bits(8K bytes) is allocated on behalf of possible VMID
range, which is two pages for some architectures. Besides that, its memory
consumption is static, despite of how many VMIDs are allocated actually.

Thats memory inefficient and lack of scalability.

So an IDA based implementation is introduced to address this issue, which
is capable of self-expanding on the actual need of VMID allocation.

Signed-off-by: Dawei Li <set_pte_at@outlook.com>
Reviewed-by: Jason Gunthorpe <jgg@nvidia.com>
Link: https://lore.kernel.org/r/TYCP286MB2323E0C525FF9F94E3B07C7ACA35A@TYCP286MB2323.JPNP286.PROD.OUTLOOK.COM
Signed-off-by: Will Deacon <will@kernel.org>

MTK-Commit-Id: 922b9f3e0b05232aca817a0bd062ebef0e29c1c1

Signed-off-by: Yunfei Wang <yf.wang@mediatek.com>
CR-Id: ALPS08609695
Feature: iommu_kernel
Change-Id: I7fdf2e2272a8ce11576136bbbb4a1eca282e93c1