[ALPS08564265] Touch: add touch ic reset operation
In QEA temperature testing, if temperature falls below the range
specified in the datasheet, touch IC may malfunction to report erratic
touch points. In some scenarios, this could lead to NE.
Adds a mechanism in ESD checking workqueue to reduce false touch issues
under low-temperature conditions. It monitors the temperature changes
and sets a flag for a deferred reset when the temperature is below zero
and the change exceeds 10 degrees. The actual reset is deferred until
the interrupt bottom half when no finger touch is detected, ensuring
uninterrupted regular touch operation. The deferred reset flag is
cleared when suspend/resume as the reset or power off/on is included in
suspend/resume operation.

MTK-Commit-Id: 2ac0e61effb5df701ab6ad07b65cd09042f1f4e9

Signed-off-by: mtk27743 <jianming.liu@mediatek.com>
CR-Id: ALPS08564265
Feature: touch_driver
Change-Id: I13bec934303a1646b4cecf2d4305240da42bd6cb