[ALPS08707003] iommu: flush IOTLB only if domain has been attached
If an IOMMU domain was never attached, it lacks any linkage to
the actual IOMMU hardware. Attempting to do flush_iotlb_all()
on it will result in a NULL pointer dereference. This seems
to happen after the recent IOMMU core rework in v6.4-rc1.

Unable to handle kernel read from unreadable memory at
virtual address 0000000000000018
    Call trace:
     mtk_iommu_flush_iotlb_all+0x20/0x80
     iommu_create_device_direct_mappings.part.0+0x13c/0x230
     iommu_setup_default_domain+0x29c/0x4d0
     iommu_probe_device+0x12c/0x190
     of_iommu_configure+0x140/0x208
     of_dma_configure_id+0x19c/0x3c0
     platform_dma_configure+0x38/0x88
     really_probe+0x78/0x2c0

Check if the "bank" field has been filled in before actually
attempting the IOTLB flush to avoid it. The IOTLB is also
flushed when the device comes out of runtime suspend, so it
should have a clean initial state.

MTK-Commit-Id: 150f039862c13159e726c6c484ccf4ba96422bdf

CR-Id: ALPS08707003
Signed-off-by: Qiming Sun <qiming.sun@mediatek.com>
Feature: iommu_kernel
Change-Id: I924f4856ed4a0b63cda78ebac1be9cdf6c223015