[ALPS08564265] arm64: dts: mt6991: set irq flag property
In newer versions of the linux kernel, the of_get_named_gpio_flags
function was removed. The irq flag is theoretically set in the DTS,
but for compatibility with previous chips, the driver uses a fixed
irq-flags instead of the one from the DTS. However, modifications are
still made in the DTS to maintain the correct values for logic trace.

MTK-Commit-Id: 577c1029c3dbc03a8690d0933f5d4ce97451418d

Signed-off-by: mtk27743 <jianming.liu@mediatek.com>
CR-Id: ALPS08564265
Feature: touch_driver
Change-Id: I4b20742de039da6681fd7585be01bac365ebc4ff