[ALPS08659114] xhci: sync upstream patch 2024_02_18-3
[Description]
xhci: print warning when HCE was set

When HCE(Host Controller Error) is set, it means that the xhci hardware
controller has an error at this time, but the current xhci driver
software does not log this event.

By adding an HCE event detection in the xhci interrupt processing
interface, a warning log is output to the system, which is convenient
for system device status tracking.

Signed-off-by: Mathias Nyman <mathias.nyman@linux.intel.com>
Signed-off-by: Greg Kroah-Hartman <gregkh@linuxfoundation.org>

[Test]
USB Host

MTK-Commit-Id: df27a6a559998ec3a3749f8de071e18f2c3f8af8

Signed-off-by: yu-chen.liu <yu-chen.liu@mediatek.com>
CR-Id: ALPS08659114
Feature: usb_ssusb
Change-Id: I8c236410b5704eadbfbf5e22eb67ddcc46c521e1