[ALPS08715211] net: stmmac: give null to dangling pointer
[Description]
When calling destroy_workqueue(), although the work queue is
destroyed,the pointer still points to the original memory address.
This creates the problem of dangling pointers, which are pointers
that point to memory that has been freed.
Attempting to access memory through a dangling pointer usually results
in undefined behavior.

[Test]
local build and test pass.

MTK-Commit-Id: 99830e38ef85cdcda8efd505ddbe3bfb58183540

Signed-off-by: larry <larry.li@mediatek.com>
CR-Id: ALPS08715211
Feature: ethernet_dwmac
Change-Id: I5293ff934f900ce7b87c6f6b9f1680a2ef07ea00