[ALPS08709396] iommu: clean up release_device checks
[Description]
Since .release_device is now called through per-device ops,
any call which gets as far as a driver definitely *is* for
that driver, for a device which has successfully passed
.probe_device, so all the checks tothat effect are now redundant
and can be removed. In the same vein we can also skip freeing
fwspecs which are now managed by core code.

Signed-off-by: Robin Murphy <robin.murphy@arm.com>
Reviewed-by: Lu Baolu <baolu.lu@linux.intel.com>
Link: https://lore.kernel.org/r/02671dbfad7a3343fc25a44222350ef
cb455fe3c.1655822151.git.robin.murphy@arm.com
Signed-off-by: Joerg Roedel <jroedel@suse.de>

[Test]
Turn the screen on and off to check whether the release dev flow
can be completed normally.

MTK-Commit-Id: 16015d86f87bb421fd6050c47ba10e87c597c598

CR-Id: ALPS08709396
Signed-off-by: Qiming Sun <qiming.sun@mediatek.com>
Feature: iommu_kernel
Change-Id: I7dfc5e99f835f757a8ff1d1d6ea6c12b9db07249