[ALPS08712363] arm64: dts: mt6893: modify dma-ranges
[Description]
Kernel6.6 obtains the dma-ranges attribute logic adjustment,
which is obtained from the parent node by default, and the
dma-ranges configured on this node are invalid. Therefore,
it is necessary to move the nodes whose dma-ranges attribute
exceeds 4G to the soc, and configure the soc dma-ranges
to be 0~16G.

[Test]
iommu basic test pass.

MTK-Commit-Id: b28ed14a0cea84047bbc3087855fcec55bece287

Change-Id: Id62f02d1cfc5bb23478edd1421e649bd44713bb3
CR-Id: ALPS08712363
Signed-off-by: "ot_vikram.b" <ot_vikram.b@mediatek.com>
Feature: iommu_kernel