[ALPS08691756] arm64: dts: mt6765: modify dma-ranges
[Description]
Kernel6.6 obtains the dma-ranges attribute logic adjustment,
which is obtained from the parent node by default, and the
dma-ranges configured on this node are invalid. Therefore,
it is necessary to move the nodes whose dma-ranges attribute
exceeds 4G to the soc, and configure the soc dma-ranges
to be 0~16G.

[Test]
iommu basic test pass.

MTK-Commit-Id: 3fd785135191c250d2778d213e196dcf2bb26284

Change-Id: I98ffdf12a378b8c197762642228b1b895f2baeff
CR-Id: ALPS08691756
Signed-off-by: "ot_vikram.b" <ot_vikram.b@mediatek.com>
Feature: iommu_kernel