[ALPS08853062] irq_mon: add showacpu to kp list
[Description]
showacpu() is a function designed for debugging purposes triggered by
the native `sysrq`, which dumps the CPU stack within the IRQ context.
When UART is active, this action can lead to prolonged IRQ latency. As
it is a debug behavior triggered natively, do not classify it as an IRQ
LONG issue.

[Test]
PDAS

MTK-Commit-Id: c46626eaa87e21cfc7fb905761ad67ff9af4bccf

Signed-off-by: Cheng-Jui Wang <cheng-jui.wang@mediatek.com>
CR-Id: ALPS08853062
Feature: IRQ Monitor
Change-Id: I7bdc36aa00b656ddfe6d3ce77ca9c52f1415fd42