[ALPS08659114] xhci: sync upstream patch 2024/05/20-3
[Description]
FROMLIST: usb: host: xhci: Repurpose event handler for
skipping interrupter events

Depending on the interrupter use case, the OS may only
be used to handle the interrupter event ring clean up.
In these scenarios, event TRBs dont need to be handled
by the OS, so introduce an xhci interrupter flag to tag
if the events from an interrupter needs to be handled
or not.
[Test]
usb host

MTK-Commit-Id: 0b6b22f8cc1d77612e54fec15b15b24956332304

Signed-off-by: yu-chen.liu <yu-chen.liu@mediatek.com>
CR-Id: ALPS08659114
Feature: usb_ssusb
Change-Id: Ib16431a33431328f10b4bcc549d2219204440d1d