[ALPS08997856] mm: fix vbq->free list corruption
[Description]
A vb might be in the free list of vbq A, but in the xarray
of vbq B, so a vb scanned from xrrary of vbq B might belong
to the free list of vbq A, remove the vb from the free list
of vbq A needs to take the lock of vbq A instead of vbq B.

Scan free list to free vb to make sure the right lock is took.

[Test]
vm_map_ram() stress.

MTK-Commit-Id: c9f6b998fc3993a979ff07a0e7b5ae65b05e88a3

Signed-off-by: Andrew Yang <andrew.yang@mediatek.com>
CR-Id: ALPS08997856
Feature: Kernel Memory Management
Change-Id: I8251ec26f19ff717863af6799510012ce9db474c