[ALPS09112047] irq_mon: irq_log support for recording function entry/exit
[Description]
Currently, irq_log_store needs to be embedded within a function to
record time expenses. This method is not applicable when dealing with
functions that cannot be directly modified or when there is a need to
record the execution time of callback functions. This is especially true
for irq monitor tracers, which are mostly the latter case. To facilitate
the integration of irq monitor tracers into irq_log, support for
recording function entry and function exit has been added.

[Test]
boot pass and PDAS

MTK-Commit-Id: 325b98b13d2ae2c93be1f1afe22512dcfce70ec7

Signed-off-by: Cheng-Jui Wang <cheng-jui.wang@mediatek.com>
CR-Id: ALPS09112047
Feature: IRQ Monitor
Change-Id: I260abbc66c15097328312ccb417aba08552d8977