[ALPS09070892] KPM: xhci: Add a quirk for writing ERST in high-low order
[Description]
This quirk is for the controller that has a limitation in supporting
separate ERSTBA_HI and ERSTBA_LO programming. Its supported when
the ERSTBA is programmed ERSTBA_HI before ERSTBA_LO. Thats because
the internal initialization of event ring fetches the
"Event Ring Segment Table Entry" based on the indication of ERSTBA_LO
written.

[Test]
usb host

MTK-Commit-Id: 9d90d1497530d210a64e73dcd6e14839f4a23054

Signed-off-by: yu-chen.liu <yu-chen.liu@mediatek.com>
CR-Id: ALPS09070892
Feature: usb_ssusb
Change-Id: I24d4d479aadb18786f2df97389288855fe32f23d