[ALPS09070892] xhci: mbrain: fix the un-changed state case
[Description]
if some device keeps in a state over 2s, we assume it is in the final
state (enumeration ok or failed).
However for the enumeration failed case, xhci won’t send the urb again
(xhci_trace_ep_urb_giveback() won’t be invoked). And won’t update the
mbrain db. To fix that, queue a work to handle it.

[Test]
insert an usb device and update mbrain db well

MTK-Commit-Id: e74f234b115173cc94ccabffc67e0d4c403c13a2

Change-Id: I6a98a1ebac96f07f95239086047d041a1814092d
Signed-off-by: Denis Hsu <denis.hsu@mediatek.com>
CR-Id: ALPS09070892
Feature: usb_ssusb