[ALPS08634965] blocktag: fix btag printf for null case
[Description]
BTAG_PRINTF() macro generalizes the usage of seq_file, char buff and
pr_info. When both buff and seq_file are null, pr_info() is expected
to be used for logging. However, directly passing NULL would cause a
build error when the macro is expanded. To avoid this issue, declare
new variables in the macro and assigned to the input parameters.

[Test]
bootup pass

MTK-Commit-Id: cf8e285b83904731b2411f898b2f0ff71515d688

Signed-off-by: Ed Tsai <ed.tsai@mediatek.com>
CR-Id: ALPS08634965
Feature: blocktag
Change-Id: Ia0478e3ba606ae9969fa109fdb68b4adda903c3e