[ALPS09112047] irq_mon: Refactor burst IRQ check logic
[Description]
This change is a logical optimization and does not affect execution
results.

- Simplified parameters and logic of `check_burst_irq`
- Introduced `in_plist` function to check if an IRQ name is in the
  predefined list.
- Moved t_avg calculation to `irq_count_core` as `in_plist` also needs
  it
- Relocated `check_burst_irq` before acquiring the spinlock
- Replaced manual spinlock handling with `scoped_guard` for better
  readability and safety.
- Moved `aee_limit` and debouce check closer to the AEE API

This change moves operations that do not require `irq_desc` information
outside of `desc->lock`. It is now more likely to return before
acquiring the spinlock, reducing the number of times the spinlock is
touched.
This modification is also in preparation for upcoming changes.

[Test]
PDAS

MTK-Commit-Id: f5faf802e5a5334fb3ea395742f609e750ab5a33

Signed-off-by: Cheng-Jui Wang <cheng-jui.wang@mediatek.com>
CR-Id: ALPS09112047
Feature: IRQ Monitor
Change-Id: I82569d05dd6cdc956dce8712a025994e12ee0d88