[ALPS08958911] SPI:add infra_req
[Description]
background:
      On the new platform, the infra fast on/off feature will cause
the mux vote box to be cleared unexpectedly, causing the pll to
disappear during spi work. Therefore, infra should be requested
during the operation of the ip.

solution:
1. Doing request and release operations for infra on under the RPM
architecture.
2. There is no dependency between the infra operation and the ccf
operation in clk, so the two must be decoupled during the rpm operation.
3. When requesting, a retry operation is required because hardware
resources are involved. When releasing, we only need to use software
to set the flag, so there is no need to retry.
4. All controllers using one infra request in rpm will cause a race
condition, so one controller corresponds to one infra request.
5. The architecture of rpm has ensured that the operation of infra is
serial, so there is no need for lock protection when operating a
single controller.
6, When infra request not success,directly BUG_ON
7, Because the controller index is needed in the probe,the patch
increases the compatibility with previous platforms.
8, Add some error handle,when clk option fail.
[Test]
Use multiple controllers to SPI UT stress test

MTK-Commit-Id: 44a5db0e01d0cd74d704866ce9e4fec5b3807bce

Signed-off-by: yijian jia <yijian.jia@mediatek.com>
CR-Id: ALPS08958911
Feature: spi_master
Change-Id: I15da4955ac10e478739d2627256f105f86786311