[ALPS09521898] mt6899: disp: add enum for preurgent to select mode
[Description]
Add an opt for pre-urgent to select to be enabled in which disp mode,
to avoid another display mode being mis-enabled.
This is because the functions are different in different modes:
In vdo mode is to avoid underrun, caused by insufficient prefetch.
In cmd mode is to avoid frame drop, caused by too long prefetch time.

[Test]
Display Stress.

MTK-Commit-Id: 4e0bf9af3709267fa8abf7e985307ecbd471d637

Signed-off-by: kai.shi <kai.shi@mediatek.com>
CR-Id: ALPS09521898
Feature: display_mobile
Change-Id: I6cecf953e9235ab0151d7c6cbce1cfe5da803f05