[ALPS09490234] rndis: sync upstream patch 2024/10/25-2
[Description]
Revert "usb: gadget: u_ether: Do not make UDC parent of the net device"

This reverts commit 321b59870f850a10dbb211ecd2bd87b41497ea6f.

This commit broke USB networking on Ingenic SoCs and maybe elsewhere.
The actual reason is unknown; and while a proper fix would be better,
were sitting at -rc7 now, so a revert is justified - and we can work on
re-introducing this change for 6.3.

Fixes: ("usb: gadget: u_ether: Do not make UDC parent of the net device")

[Test]
USB RNDIS test

MTK-Commit-Id: ee84a7827a94fe67d877a072c5a599f1cd7139d3

Signed-off-by: Paul Cercueil <paul@crapouillou.net>
Acked-by: Sascha Hauer <s.hauer@pengutronix.de>
Link: https://lore.kernel.org/r/20230209105626.10597-1-paul@crapouillou.net
Signed-off-by: Greg Kroah-Hartman <gregkh@linuxfoundation.org>
Change-Id: If8bb773e9d8157dec56fcd4cd9070327911f8c75
Signed-off-by: Chih-Chun Wang <Chih-Chun.Wang@mediatek.com>
CR-Id: ALPS09490234
Feature: usb_ssusb