[ALPS09143903] Runtime determination of counter frequency and width
[Description]
This patch introduces runtime determination of whether to use a 1GHz
or 13MHz counter for timesync in SSPM driver. The changes include:

1. Determine the appropriate counter (1GHz or 13MHz) at runtime and
   compute the suitable width.
2. Modify the timesync configuration to support runtime determination,
   ensuring that the 1GHz counter, which requires a 61-bit mask, is
   handled correctly.

The original static configuration could not support the 1GHz counter
properly and would break it due to insufficient width. By dynamically
determining the counter frequency, we ensure that both 1GHz and legacy
13MHz counters are supported.

[Test]
Build pass

MTK-Commit-Id: f287e562c73886ebbbc3a1e30600c32b1b530e69

Signed-off-by: Walter Chang <walter.chang@mediatek.com>
CR-Id: ALPS09143903
Feature: RGU WDT Watchdog
Change-Id: Ia2aa8b91f1d41e42131b80bc39313a869dad3bfd