[ALPS09162427] smmu/mediatek: Fix last_sid_idx calculation
[Description]
iommu/arm-smmu-v3: Fix last_sid_idx calculation for sid_bits==32

The function arm_smmu_init_strtab_2lvl uses the expression

((1 << smmu->sid_bits) - 1)

to calculate the largest StreamID value. However, this fails for the
maximum allowed value of SMMU_IDR1.SIDSIZE which is 32. The C standard
states:

"If the value of the right operand is negative or is greater than or
equal to the width of the promoted left operand, the behavior is
undefined."

With smmu->sid_bits being 32, the prerequisites for undefined behavior
are met.  We observed that the value of (1 << 32) is 1 and not 0 as we
initially expected.

Similar bit shift operations in arm_smmu_init_strtab_linear seem to not
be affected, because it appears to be unlikely for an SMMU to have
SMMU_IDR1.SIDSIZE set to 32 but then not support 2-level Stream tables

This issue was found by Ryan Huang <tzukui@google.com> on our team.

Fixes: ce410410f1a7 ("iommu/arm-smmu-v3: Add arm_smmu_strtab_l1/2_idx()")
Signed-off-by: Daniel Mentz <danielmentz@google.com>

[Test]
Build pass and pass self-ktf test.

MTK-Commit-Id: 51685cbf6387b69b31486454324ae373232be392

Signed-off-by: mingyuan ma <mingyuan.ma@mediatek.com>
CR-Id: ALPS09162427
Feature: iommu_kernel
Change-Id: I6bf10455a2645ff4c4eb60e8afe5a8bc8156023b