[ALPS09666842] Fix direction flag assignment in MMC blocktag
[Description]
Correct the assignment of the direction flag in the
`mmc_mtk_biolog_send_command` function. The previous implementation
used a double negation (!!) to determine the direction, which has been
replaced with a single negation (!). This change ensures the correct
direction flag is set for both SD and non-SD cases when processing MMC
requests.

- Updated direction flag assignment for SD and non-SD cases.
- Ensured proper calculation of the direction flag based on the
MMC_DATA_READ flag.

This fix addresses the issue of confirming the value of the
/proc/blocktag/mmc/blockio node in the kernel-6.6.

[Test]
Bootup and verify pass

MTK-Commit-Id: 1855a7653fdccdfb03f06e873de022f7935119dd

CR-Id: ALPS09666842
Feature: blocktag
Change-Id: Ia3adc08e9625b43c3c51707efa928c9e7a6771dc
Signed-off-by: yanxu <yanxu.wei@mediatek.com>