[ALPS09563159] smmu/mediatek: Update comments about ATS and bypass
[Description]
iommu/arm-smmuv3: Update comments about ATS and bypass

The SMMUv3 spec has a note that BYPASS and ATS dont work together
under the STE EATS field definition. However there is another
section "13.6.4 Full ATS skipping stage 1" that explains under
certain conditions BYPASS and ATS do work together if the STE is
using S1DSS to select BYPASS and the CD table has the possibility
for a substream.

When these comments were written the understanding was that all
forms of BYPASS just didnt work and this was to be a future
problem to solve.

It turns out that ATS and IDENTITY will always work just fine:

 - If STE.Config = BYPASS then the PCI ATS is disabled

 - If a PASID domain is attached then S1DSS = BYPASS and ATS will be
   enabled. This meets the requirements of 13.6.4 to automatically
   generate 1:1 ATS replies on the RID.

Update the comments to reflect this.

Fixes: "iommu/arm-smmu-v3: Make changing domains be hitless for ATS"
Signed-off-by: Jason Gunthorpe <jgg@nvidia.com>
Link: https://lore.kernel.org/r/0-v1-f27174f44f39+27a33-smmuv3_ats_note_jgg@nvidia.com
Signed-off-by: Will Deacon <will@kernel.org>
Signed-off-by: Sasha Levin <sashal@kernel.org>

[Test]
Build pass and pass self-ktf test.

MTK-Commit-Id: 44160006f0c42433ffc9bf32b0a979d725794009

Signed-off-by: mingyuan ma <mingyuan.ma@mediatek.com>
CR-Id: ALPS09563159
Feature: iommu_kernel
Change-Id: Ida6b867bc7e0abb7eeb9370b2b9dc7fcf983364c