[ALPS08659114] xhci: sync upstream patch of.p200-1
[Description]
BACKPORT: xhci: add helper to stop endpoint and wait for completion

Expose xhci_stop_endpoint_sync() which is a synchronous variant of
xhci_queue_stop_endpoint().  This is useful for client drivers that are
using the secondary interrupters, and need to stop the current endpoint
session.

This does not go through the normal xhci_handle_cmd_stop_ep() command
completion handler, because it utilizes the completion path to achieve
synchronous behavior.  Users of this API are primarily intended to be
clients that maintain their own transfer rings, such as in the case of USB
audio offload.
[Test]
USB Audio Auto Test

MTK-Commit-Id: 3da873679ca8fdf22b9cd98b979459b60b3a6e6f

Signed-off-by: yu-chen.liu <yu-chen.liu@mediatek.com>
CR-Id: ALPS08659114
Feature: usb_ssusb
Change-Id: Ife2462205e3d36aabc8aeca709d2ad659934b595