[ALPS08734212] venc add timing info option for MTK encoder
[Description]
Add c2 key to configure the MTK encoder to include timing
information in the encoded bitstream. This timing information
is typically included in the VUI parameters of the SPS in the
H.264 bitstream, and in the VUI parameters of the VPS in the H.265
bitstream.

c2 key mtk.venc.encoding.option.encoded-with-timing-info:
- If set to 1, enables the inclusion of timing information such as
num_units_in_tick, time_scale, and fixed_frame_rate_flag in the VUI
parameters of the SPS for H.264, and in the VUI parameters of the
VPS for H.265.

[Test]
tc_030

MTK-Commit-Id: a23e5853070ae3f582b8d003625dcc0749fe886d

CR-Id: ALPS08734212
Feature: Video Driver
Signed-off-by: bernie.yu <Bernie.Yu@mediatek.com>
Change-Id: I526ebfd752b0a743d70e1209b6d13cdd07894bcd