[ALPS09967707] Fix [Coverity][ALPS]SEC_UNCHECK_SUBTRACTION: 11506033
[Description]
The issue at line 360 involves subtracting an unsigned variable, which
can cause an overrun issue if the result is negative. The original
code snippet checks if alloc - get is greater than log->sz, but if
alloc is less than get, the result would be a large positive number
due to unsigned integer wrap-around, leading to incorrect behavior.

To fix this, Ive modified the condition in the if statement to
if (alloc > get + log->sz). This ensures that the subtraction is
checked in a way that prevents wrap-around issues. Specifically, by
checking if alloc is greater than get + log->sz, we ensure the
subtraction operation will not result in a negative value that
wraps around to a large positive value.

This approach guarantees that the subtraction performed in the
assignment get = alloc - log->sz is safe and does not cause an
overflow, as the condition now ensures that alloc is sufficiently
larger than get + log->sz. Thus, the subtraction result will always be
within a valid range for unsigned integers.

[Test]
Build pass, boot pass

MTK-Commit-Id: a9557c8cc377eb08cc0947919bb85678ebfe35a2

CR-Id: ALPS09967707
Feature: GenieZone
Signed-off-by: Liju-clr Chen <liju-clr.chen@mediatek.com>
Change-Id: I0eb12e1b31db8c4fd1eeef9a28868afbf24a99ec