[ALPS09546955] ovl: add matrix property
[Description]
Add a new plane property "OVL_Y2R_MATRIX_IDX" to get Y2R configuration
from userspace. If the property value is Y2R_MATRIX_LEGACY_MODE, keep
to use the original logic. If it is Y2R_MATRIX_DISABLE, we disable the
the Y2R. On the other hand, we try to use the configuration as Y2R
matrix.

[Test]
Check the composited result whether DPU is the same with GPU or not.

MTK-Commit-Id: 0ad9b225da997d6322e218ebd2559f55f9f83322

CR-Id: ALPS09546955
Feature: display_mobile
Change-Id: I8622f8414f218001f59ccb5720520250ecc341fc
Signed-off-by: Aaron Wu <ycw.wu@mediatek.com>