[ALPS10158508] ASoC: Read DTS speaker settings in bypass mode
[Description]
Resolved an issue where an assertion was triggered in HAL when booting
the EVB with a phone load. The issue was caused by the misuse of an
undefined I2S interface, as EVB is not supposed to use the phone load,
this issue occurs as expected.

Even in bypass mode, DTS speaker settings are still read to ensure
maximum compatibility with different loads.

[Test]
Is able to read DTS speaker settings with EVB + phone load.

MTK-Commit-Id: 80a971a2ce481d209dbdc8cc58004d3da91e2ccc

CR-Id: ALPS10158508
Feature: alsa_sound_soc
Signed-off-by: Shawn Sung <shawn.sung@mediatek.com>
Change-Id: I7277dbdbfa7cf7811bba5f4c5a7e464121589320