[ALPS10509767] mediatek: vhost_dmabuf_phys: Update to latest
[Description]
1. Add hw_clk support
The SW clock can not help to synchronize for the apps inside both OS.
It always has a base offset. thus HW-clk is expected.
hw-clk has nothing to do with dmabuf. It is added here just for the
convenience of dmabuf users.
2. Add cache invalid and flush for vhost_dmabuf_exporter_phys device
node. if not, there will be a flash screen since the camera
buffer is not cached-sync.
3.Adapt sec dmabuf map when SECURE_MODE setting.
Skip CPU cache sync in sec mode, so the sgtable information passed
by the guest cannot be converted by the host to the CPU VA for
access, skip is required in dma_map.
4.delete "enmu mode", use "bool is_secure" instead

[Test]
build pass

MTK-Commit-Id: 64cbfef43f9492ec407cc4e271fef7a94b9a1512

Signed-off-by: jiaqi sun <ot_jiaqi.sun@mediatek.com>
CR-Id: ALPS10509767
Feature: vdmabuf
Change-Id: Ieb4f41044d6cda6d828c841d5412623a8c6c1f8c
(cherry picked from commit 79b78fac2bc038ef0dbdb25414becd4e7bb3ba9c)